inference/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# PyTorch and ONNX
torch==2.1.2
torchvision==0.16.2
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3

# TorchXRayVision for CXR classification
torchxrayvision==1.2.1

# Image processing
Pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
scikit-image==0.22.0
scipy==1.12.0

# Calibration
scikit-learn==1.4.0

# Utilities
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
